[CARBONDATA-3889] Enable java check style for all java modules
Why is this PR needed?
some java modules skip to check java style
There are java files under the scala source directory

What changes were proposed in this PR?
enable check style for all java modules
Move java files into the java source directory

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3872
