+++ core/src/main/java/org/apache/carbondata/core/metadata/schema/indextable/IndexTableInfo.java
@@ -176,6 +176,4 @@ public class IndexTableInfo implements Serializable {
   public void setIndexProperties(Map<String, String> indexProperties) {
     this.indexProperties = indexProperties;
   }
-
-
 }
