+++ integration/presto/src/main/prestosql/org/apache/carbondata/presto/readers/ByteStreamReader.java
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.carbondata.presto.readers;
 
 import org.apache.carbondata.core.metadata.datatype.DataType;
