+++ integration/spark/src/main/java/org/apache/spark/sql/ColumnVectorFactory.java
@@ -18,26 +18,25 @@
 package org.apache.spark.sql;
 
 import org.apache.spark.memory.MemoryMode;
-import org.apache.spark.sql.execution.vectorized.OnHeapColumnVector;
 import org.apache.spark.sql.execution.vectorized.OffHeapColumnVector;
+import org.apache.spark.sql.execution.vectorized.OnHeapColumnVector;
 import org.apache.spark.sql.execution.vectorized.WritableColumnVector;
 import org.apache.spark.sql.types.StructType;
 
 public class ColumnVectorFactory {
 
-    public static WritableColumnVector[] getColumnVector(MemoryMode memMode, StructType outputSchema, int rowNums) {
+  public static WritableColumnVector[] getColumnVector(MemoryMode memMode, StructType outputSchema,
+      int rowNums) {
 
-        WritableColumnVector[] writableColumnVectors = null;
-        switch (memMode) {
-            case ON_HEAP:
-                writableColumnVectors = OnHeapColumnVector
-                        .allocateColumns(rowNums, outputSchema);
-                break;
-            case OFF_HEAP:
-                writableColumnVectors = OffHeapColumnVector
-                        .allocateColumns(rowNums, outputSchema);
-                break;
-        }
-        return writableColumnVectors;
+    WritableColumnVector[] writableColumnVectors = null;
+    switch (memMode) {
+      case ON_HEAP:
+        writableColumnVectors = OnHeapColumnVector.allocateColumns(rowNums, outputSchema);
+        break;
+      case OFF_HEAP:
+        writableColumnVectors = OffHeapColumnVector.allocateColumns(rowNums, outputSchema);
+        break;
     }
+    return writableColumnVectors;
+  }
 }
