+++ integration/spark/src/main/java/org/apache/spark/sql/secondaryindex/load/CarbonInternalLoaderUtil.java
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.spark.sql.secondaryindex.load;
 
 import java.io.IOException;
@@ -36,7 +37,6 @@ import org.apache.carbondata.core.statusmanager.LoadMetadataDetails;
 import org.apache.carbondata.core.statusmanager.SegmentStatus;
 import org.apache.carbondata.core.statusmanager.SegmentStatusManager;
 import org.apache.carbondata.core.util.path.CarbonTablePath;
-import org.apache.carbondata.processing.loading.model.CarbonLoadModel;
 import org.apache.carbondata.processing.util.CarbonLoaderUtil;
 
 import org.apache.log4j.Logger;
@@ -44,12 +44,11 @@ import org.apache.spark.sql.index.CarbonIndexUtil;
 
 public class CarbonInternalLoaderUtil {
 
-  private static final Logger LOGGER =
-      LogServiceFactory.getLogService(CarbonInternalLoaderUtil.class.getName());
+  private static final Logger LOGGER = LogServiceFactory.getLogService(
+      CarbonInternalLoaderUtil.class.getName());
 
   public static List<String> getListOfValidSlices(LoadMetadataDetails[] details) {
-    List<String> activeSlices =
-        new ArrayList<>(CarbonCommonConstants.DEFAULT_COLLECTION_SIZE);
+    List<String> activeSlices = new ArrayList<>(CarbonCommonConstants.DEFAULT_COLLECTION_SIZE);
     for (LoadMetadataDetails oneLoad : details) {
       if (SegmentStatus.SUCCESS.equals(oneLoad.getSegmentStatus())
           || SegmentStatus.LOAD_PARTIAL_SUCCESS.equals(oneLoad.getSegmentStatus())
@@ -62,7 +61,6 @@ public class CarbonInternalLoaderUtil {
 
   /**
    * This method will return the mapping of valid segments to segment load start time
-   *
    */
   public static Map<String, Long> getSegmentToLoadStartTimeMapping(LoadMetadataDetails[] details) {
     Map<String, Long> segmentToLoadStartTimeMap = new HashMap<>(details.length);
@@ -100,18 +98,18 @@ public class CarbonInternalLoaderUtil {
           return false;
         }
 
-        LoadMetadataDetails[] currentLoadMetadataDetails =
-            SegmentStatusManager.readLoadMetadata(metaDataFilepath);
+        LoadMetadataDetails[] currentLoadMetadataDetails = SegmentStatusManager.readLoadMetadata(
+            metaDataFilepath);
 
-        List<LoadMetadataDetails> updatedLoadMetadataDetails =
-            new ArrayList<>(CarbonCommonConstants.DEFAULT_COLLECTION_SIZE);
+        List<LoadMetadataDetails> updatedLoadMetadataDetails = new ArrayList<>(
+            CarbonCommonConstants.DEFAULT_COLLECTION_SIZE);
 
         // check which load needs to be overwritten which are in in progress state
         boolean found = false;
         for (int i = 0; i < currentLoadMetadataDetails.length; i++) {
           for (LoadMetadataDetails newLoadMetadataDetail : newLoadMetadataDetails) {
-            if (currentLoadMetadataDetails[i].getLoadName()
-                .equals(newLoadMetadataDetail.getLoadName())) {
+            if (currentLoadMetadataDetails[i].getLoadName().equals(
+                newLoadMetadataDetail.getLoadName())) {
               currentLoadMetadataDetails[i] = newLoadMetadataDetail;
               found = true;
               break;
@@ -126,7 +124,8 @@ public class CarbonInternalLoaderUtil {
         for (int i = 0; i < newLoadMetadataDetails.size(); i++) {
           foundNext = false;
           for (int j = 0; j < currentLoadMetadataDetails.length; j++) {
-            if (newLoadMetadataDetails.get(i).getLoadName().equals(currentLoadMetadataDetails[j].getLoadName())) {
+            if (newLoadMetadataDetails.get(i).getLoadName().equals(
+                currentLoadMetadataDetails[j].getLoadName())) {
               foundNext = true;
               break;
             }
@@ -145,28 +144,27 @@ public class CarbonInternalLoaderUtil {
 
         List<String> indexTables = CarbonIndexUtil.getSecondaryIndexes(carbonTable);
         if (!indexTables.isEmpty()) {
-          List<LoadMetadataDetails> newSegmentDetailsListForIndexTable =
-              new ArrayList<>(validSegments.size());
+          List<LoadMetadataDetails> newSegmentDetailsListForIndexTable = new ArrayList<>(
+              validSegments.size());
           for (String segmentId : validSegments) {
             LoadMetadataDetails newSegmentDetailsObject = new LoadMetadataDetails();
             newSegmentDetailsObject.setLoadName(segmentId);
             newSegmentDetailsListForIndexTable.add(newSegmentDetailsObject);
           }
           for (CarbonTable indexTable : indexCarbonTables) {
-            List<LoadMetadataDetails> indexTableDetailsList = CarbonIndexUtil
-                .getTableStatusDetailsForIndexTable(updatedLoadMetadataDetails, indexTable,
-                    newSegmentDetailsListForIndexTable);
+            List<LoadMetadataDetails> indexTableDetailsList =
+                CarbonIndexUtil.getTableStatusDetailsForIndexTable(updatedLoadMetadataDetails,
+                    indexTable, newSegmentDetailsListForIndexTable);
 
             SegmentStatusManager.writeLoadDetailsIntoFile(
                 CarbonTablePath.getTableStatusFilePath(indexTable.getTablePath()),
-                indexTableDetailsList
-                    .toArray(new LoadMetadataDetails[0]));
+                indexTableDetailsList.toArray(new LoadMetadataDetails[0]));
           }
         } else if (carbonTable.isIndexTable()) {
           SegmentStatusManager.writeLoadDetailsIntoFile(
               metaDataFilepath + CarbonCommonConstants.FILE_SEPARATOR
-                  + CarbonTablePath.TABLE_STATUS_FILE, updatedLoadMetadataDetails
-                  .toArray(new LoadMetadataDetails[0]));
+                  + CarbonTablePath.TABLE_STATUS_FILE,
+              updatedLoadMetadataDetails.toArray(new LoadMetadataDetails[0]));
         }
         status = true;
       } else {
@@ -178,8 +176,7 @@ public class CarbonInternalLoaderUtil {
       LOGGER.error(
           "Not able to acquire the lock for Table status update for table " + databaseName + "."
               + tableName);
-    }
-    finally {
+    } finally {
       if (carbonLock.unlock()) {
         LOGGER.info("Table unlocked successfully after table status update" + databaseName + "."
             + tableName);
@@ -209,12 +206,12 @@ public class CarbonInternalLoaderUtil {
     }
     boolean isIndexTableSegmentsCompacted = false;
     if (null != currentIndexTable) {
-      LoadMetadataDetails[] existingLoadMetaDataDetails =
-          SegmentStatusManager.readLoadMetadata(currentIndexTable.getMetadataPath());
+      LoadMetadataDetails[] existingLoadMetaDataDetails = SegmentStatusManager.readLoadMetadata(
+          currentIndexTable.getMetadataPath());
       for (LoadMetadataDetails existingLoadMetaDataDetail : existingLoadMetaDataDetails) {
         for (LoadMetadataDetails newLoadMetadataDetail : newLoadMetadataDetails) {
-          if (existingLoadMetaDataDetail.getLoadName()
-              .equalsIgnoreCase(newLoadMetadataDetail.getLoadName())
+          if (existingLoadMetaDataDetail.getLoadName().equalsIgnoreCase(
+              newLoadMetadataDetail.getLoadName())
               && existingLoadMetaDataDetail.getSegmentStatus() == SegmentStatus.COMPACTED) {
             isIndexTableSegmentsCompacted = true;
             break;
@@ -232,7 +229,6 @@ public class CarbonInternalLoaderUtil {
 
   /**
    * method to update table status in case of IUD Update Delta Compaction.
-   *
    */
   public static boolean updateLoadMetadataWithMergeStatus(CarbonTable indexCarbonTable,
       String[] loadsToMerge, String mergedLoadNumber, Map<String, String> segmentToLoadStartTimeMap,
@@ -240,8 +236,8 @@ public class CarbonInternalLoaderUtil {
       List<String> rebuiltSegments) throws IOException {
     boolean tableStatusUpdateStatus = false;
     List<String> loadMergeList = new ArrayList<>(Arrays.asList(loadsToMerge));
-    SegmentStatusManager segmentStatusManager =
-        new SegmentStatusManager(indexCarbonTable.getAbsoluteTableIdentifier());
+    SegmentStatusManager segmentStatusManager = new SegmentStatusManager(
+        indexCarbonTable.getAbsoluteTableIdentifier());
 
     ICarbonLock carbonLock = segmentStatusManager.getTableStatusLock();
 
@@ -249,14 +245,14 @@ public class CarbonInternalLoaderUtil {
       if (carbonLock.lockWithRetries()) {
         LOGGER.info("Acquired lock for the table " + indexCarbonTable.getDatabaseName() + "."
             + indexCarbonTable.getTableName() + " for table status update ");
-        LoadMetadataDetails[] loadDetails =
-            SegmentStatusManager.readLoadMetadata(indexCarbonTable.getMetadataPath());
+        LoadMetadataDetails[] loadDetails = SegmentStatusManager.readLoadMetadata(
+            indexCarbonTable.getMetadataPath());
 
         long modificationOrDeletionTimeStamp = CarbonUpdateUtil.readCurrentTime();
         for (LoadMetadataDetails loadDetail : loadDetails) {
           // check if this segment is merged.
-          if (loadMergeList.contains(loadDetail.getLoadName()) || loadMergeList
-              .contains(loadDetail.getMergedLoadName())) {
+          if (loadMergeList.contains(loadDetail.getLoadName()) || loadMergeList.contains(
+              loadDetail.getMergedLoadName())) {
             // if the compacted load is deleted after the start of the compaction process,
             // then need to discard the compaction process and treat it as failed compaction.
             if (loadDetail.getSegmentStatus() == SegmentStatus.MARKED_FOR_DELETE) {
@@ -279,8 +275,8 @@ public class CarbonInternalLoaderUtil {
         loadMetadataDetails.setSegmentFile(SegmentFileStore.genSegmentFileName(mergedLoadNumber,
             String.valueOf(segmentToLoadStartTimeMap.get(mergedLoadNumber)))
             + CarbonTablePath.SEGMENT_EXT);
-        CarbonLoaderUtil
-            .addDataIndexSizeIntoMetaEntry(loadMetadataDetails, mergedLoadNumber, indexCarbonTable);
+        CarbonLoaderUtil.addDataIndexSizeIntoMetaEntry(loadMetadataDetails, mergedLoadNumber,
+            indexCarbonTable);
         if (rebuiltSegments.contains(loadMetadataDetails.getLoadName())) {
           loadMetadataDetails.setLoadStartTime(newLoadStartTime);
         } else {
@@ -322,12 +318,11 @@ public class CarbonInternalLoaderUtil {
   /**
    * Method to check if main table and SI have same number of valid segments or not
    */
-  public static boolean checkMainTableSegEqualToSISeg(LoadMetadataDetails[] mainTableLoadMetadataDetails,
-                                                      LoadMetadataDetails[] siTableLoadMetadataDetails) {
-    List<String> mainTableSegmentsList =
-        getListOfValidSlices(mainTableLoadMetadataDetails);
-    List<String> indexList =
-        getListOfValidSlices(siTableLoadMetadataDetails);
+  public static boolean checkMainTableSegEqualToSISeg(
+      LoadMetadataDetails[] mainTableLoadMetadataDetails,
+      LoadMetadataDetails[] siTableLoadMetadataDetails) {
+    List<String> mainTableSegmentsList = getListOfValidSlices(mainTableLoadMetadataDetails);
+    List<String> indexList = getListOfValidSlices(siTableLoadMetadataDetails);
     Collections.sort(mainTableSegmentsList);
     Collections.sort(indexList);
     // In the case when number of SI segments are more than the main table segments do nothing
@@ -365,7 +360,8 @@ public class CarbonInternalLoaderUtil {
     if (mainTableLoadMetadataDetails.length != 0) {
       for (LoadMetadataDetails loadDetail : mainTableLoadMetadataDetails) {
         // if load in progress and check if same load is present in SI.
-        if (SegmentStatusManager.isLoadInProgress(carbonTable.getAbsoluteTableIdentifier(), loadDetail.getLoadName())) {
+        if (SegmentStatusManager.isLoadInProgress(carbonTable.getAbsoluteTableIdentifier(),
+            loadDetail.getLoadName())) {
           if (!allSiSlices.contains(loadDetail.getLoadName())) {
             return false;
           }
