+++ integration/spark/src/main/java/org/apache/spark/sql/secondaryindex/load/RowComparator.java
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.spark.sql.secondaryindex.load;
 
 import java.util.Comparator;
